refactor: 상수 파일 이동
