+++ src/apis/MovieApi.ts
@@ -1,14 +1,14 @@
 import { MoviesResponse } from "../../types/movieApiType";
 
-async function fetchWithErrorHandling(url: string) {
-  const options = {
-    method: "GET",
-    headers: {
-      accept: "application/json",
-      Authorization: `Bearer ${import.meta.env.VITE_TMDB_TOKEN}`,
-    },
-  };
+const options = {
+  method: "GET",
+  headers: {
+    accept: "application/json",
+    Authorization: `Bearer ${import.meta.env.VITE_TMDB_TOKEN}`,
+  },
+};
 
+async function fetchWithErrorHandling(url: string) {
   return fetch(url, options).then((res) => {
     if (res.ok) return res.json() as unknown as MoviesResponse;
     throw new Error(String(res.status));
